requirements.txt
# Core Flask dependencies
Flask>=3.0.0
flask-cors>=6.0.0
gunicorn>=21.0.0
python-dotenv>=1.0.0

# Google Cloud & AI - Let ADK control versions
google-adk>=0.1.0
google-genai>=0.1.0
google-auth>=2.0.0
google-cloud-bigquery>=3.0.0
google-cloud-storage>=2.10.0
google-cloud-texttospeech>=2.14.0
google-generativeai>=0.3.0
google-cloud-pubsub==2.18.4
db-dtypes>=1.1.0

# Data processing
# Data Processing
pandas>=1.5.0
PyPDF2>=3.0.0
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# Export formats
openpyxl>=3.1.0
reportlab>=4.0.0
matplotlib>=3.7.0

# Utilities
requests>=2.31.0
feedparser>=6.0.0
xmltodict>=0.13.0

# US Location Data - lightweight, no complex dependencies
zipcodes>=1.2.0

# Firebase Authentication
firebase-admin>=6.0.0